Refactor TracksResults to use audius-query
