+++ packages/web/src/pages/MainContentContext.tsx
@@ -1,10 +1,21 @@
-import { createContext, memo, useRef, MutableRefObject } from 'react'
+import {
+  createContext,
+  memo,
+  useRef,
+  MutableRefObject,
+  useContext
+} from 'react'
 
 export const MAIN_CONTENT_ID = 'mainContent'
 
 export const MainContentContext = createContext({
   mainContentRef: {} as MutableRefObject<HTMLDivElement | undefined>
 })
+
+export const useMainContentRef = () => {
+  const { mainContentRef } = useContext(MainContentContext)
+  return mainContentRef
+}
 
 export const MainContentContextProvider = memo(
   (props: { children: JSX.Element }) => {
